move descriptor to separate file
